chore: Refactoring into modules
